requirements.txt
                                                                                                               
fastapi
alembic
uvicorn
langgraph
langchain-google-genai
langchain-openai
langchain-text-splitters
google-api-python-client
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
langchain
langchain-community
aiofiles
requests
streamlit
beautifulsoup4
ddgs
crawl4ai
trafilatura
openai
praw
qdrant-client
fastembed
axiom-py
python-multipart
redis[asyncio]
python-jose[cryptography]
